[from paas-cf] Configure metron_agent as a runtime addon
We want metron_agent to be installed in all the VMs to redirect the
syslog messages to the loggregator stack.

Instead of adding the job to all VMs, we configure it as an addon.

For that, some changes were required to render `cf-manifest.yml` and
`runtime-config.yml`:

 * We need to pass the terraform outputs to the runtime-config.yml
   generation, so we can pass the environment config.
 * We need to move some common config, like dropsonde_incoming_port to
   a `common/metron.yml` file
 * We need to move the definition of the cf release to
   `common/releases.yml`

rebased from https://github.com/alphagov/paas-cf/commit/2e8e31495dfe6ecd1bf0820e82e7c4419666bb2b