Add terraform code to create release buckets
To keep the blobs of our custom releases, we need to create S3 buckets.

We decided to keep these buckets on the AWS CI account, as it is the one
that will run the pipelines to compile and publish the releases.

Meanwhile we don't have a better way to manage the configuation, we
keep in this repository the related repo that we can run locally.

The terraform state will be kept as commits in this repository.