Concourse login - set default session length to 5m
Override to 48h in dev environments for convenience. This is to force us to
log in repeatedly to ci/stage/prod - to realize that we are now operating
on these environments and not dev, as concourse interfaces are identical.

@dcarley: This is the combination of two cherry-picked commits,
alphagov/paas-cf@1472f3d and alphagov/paas-cf@5012449. The preceding commit
didn't make sense in the context of paas-bootstrap because we don't
currently have a `prod` target. I had to fix merge conflicts because the
scripts and pipeline files have moved around between repos.