[from paas-cf] Pass the etcd name to metron agent addon
With bosh v258 we must pass all the required properties to the addons.
The metron_agent requires `loggregator.etcd.machines` which currently
we set as the value in `etc.advertise_urls_dns_suffix`.

Without this property, the deployment will fail with this error for all
the VMs:

   - Unable to render jobs for instance group 'ingestor_z2'. Errors are:
     - Unable to render templates for job 'metron_agent'. Errors are:
       - Error filling in template 'metron_agent.json.erb' (line 33: Can't find property '["loggregator.etcd.machines"]')

In order to be able to deploy metron agent as addon, we must set this
property. We move the definition of the value to the file
common/metron.yml and grab it in `etcd.advertise_urls_dns_suffix` and
the addon `loggregator.etc.machines`.

Note, there are plans to remove metron_agent as addon, so this commit
must be reverted by then.

rebased from https://github.com/alphagov/paas-cf/commit/a03b617bd775c4cda3db3eec6255777cc616dcd8