Enable dedicated workers for status tasks
In bosh v257 the parameter `director.enable_dedicated_status_worker`[1]
was added. When true, it reserves one bosh worker for status tasks, like
`bosh ssh` or `bosh vms`.

This allows operate and control the platform even when other tasks are
running in background.

We enable it and add an additional worker (default was 3) to keep the
number of workers for normal tasks the same.

[1] https://bosh.io/jobs/director?source=github.com/cloudfoundry/bosh#p=director.enable_dedicated_status_worker