[from paas-cf] Pass the properties to cleanup_syslog addon
From v258 of bosh, they "fixed" it and now the addons in the runtime
config DO NOT get the properties from the main manifest. In previous
versions this was working but it was not expected.

The job syslog_cleanup_forwarding uses the property
syslog_daemon_config.enable with default true. If true, it does not do the
clean up.

This job was added with bosh release v257, abnd we had set that value to false
only in the main manifest, NOT in the runtime manifest.
As we upgraded bosh just before merge this in #570 to v258, the behaviour
changed and the cleanup job did not work.

In this PR we add this config to the runtime config, allowing it to work properly.

rebased from https://github.com/alphagov/paas-cf/commit/f4e0cd483ffe33cb654603521b93b239e318ef9c