Stop hard coding the Concourse deployment name
The task for destroying the Concourse deployment hard codes the deployment
name, meaning you have to update the deployment name in multiple places
if it were to change. The deployment name is now pulled from the manifest,
which I realised was not actually being uploaded to S3, so it has
been added as a Concourse resource.