Rename the state bucket to gds-paas-${deploy_env}-state
This bucket is now going to be used by the deployed concourse as well as
the bootstrap concourse, so it should be named to reflect this purpose.

This rename will be disruptive for existing deployments, but as we
haven't yet got anything running on these, it's a good time to do this.
Later on, this will be harder to change.