Disable RDS auto minor version upgrade
We are not sure how the next upgrade will impact us and we are planning
a major upgrade anyway in #135224045. This removes the unknown for now
and we will decide if we want to re-enable it or not in that story.

Otherwise we face the risk of AWS upgrading our instances automatically
and when our terraform runs, this would happen:

```
* aws_db_instance.cf: Error modifying DB Instance colin-nov-cf:
InvalidParameterCombination: Cannot upgrade postgres from 9.4.7 to 9.4.5
  status code: 400, request id: a4a6b0d0-b716-11e6-8cf8-29bf468258f8
```