Specify ELB security policy for concourse ELB
The default behaviour if this is unspecified is for the ELBs to get the AWS
default security policy at the time that the ELB is created. This has led to
our long-lived deployments having an older security policy, which raises
warnings in Trusted Advisor etc.

This therefore updates the concourse ELB to use the latest built-in AWS policy
(2016-08).