Fix specification of ELB security policy.
The default behaviour if this is unspecified is for the ELBs to get the
AWS default security policy at the time that the ELB is created. This
has led to our long-lived deployments having an older security policy,
which raises warnings in Trusted Advisor etc.

The previous approach only works for ELB that already have the relevant
policy defined. This was due to a misunderstanding about how these
policies work. An ELB has one or more policies defined for it. Each SSL
listener then has one of these policies attached to it. The previous
approach failed because we were attempting to attach a policy to the
listeners that didn't exist on the ELB.

This therefore changes to use a resource that defined both the policy
and the listener attachment. We can use a built-in policy by creating a
policy with the `Reference-Security-Policy` attribute set to the
relevant built-in name.