Update Bosh RDS postgres version in terraform.
This has auto_minor_version_upgrade set to true (we're not specifying it
in the terraform config, and this is the default). This instance has
therefore been upgraded automatically to 9.4.7. As a result, terraform
fail with the error:

Error modifying DB Instance build-bosh: InvalidParameterCombination:
Cannot upgrade postgres from 9.4.7 to 9.4.5

Updating this version to match will unblock the pipeline.