Bump BOSH AWS CPI to latest (v62)
Currently latest BOSH AWS CPI is v62. Bump from previous v60 adds some
fixes, but mainly makes boot/root disk volume type gp2 and not magnetic,
which means faster VM deploys and updates.

This also removes the pinning of bosh-init container. We want this to
pick up the latest version of this container as it has the correct
version of the CPI cached.