Allow STS creds to be used to bootstrap
We need to allow short-lived credentials to be used in order
to require MFA for aws users.

The sts-generated session token needs to be passed to vagrant.

The concourse atc password can no longer be generated from
aws creds, since they are no longer static but may change each day
when short-lived credentials are regenerated.
For the bootstrap pipeline, we can use a static atc password, since the
only access is via ssh - the concourse is not exposed to the
public internet. It is based on the AWS UserId so it is not
predictable easily and is unique to each user.