migrate: Handle renaming of the vpc resource.
In paas-cf this resource was called "myvpc". In this repo it's called
"paas". This is purely an internal terraform concern and doesn't affect
anything visible in AWS, but because the resource names are different,
terraform can't link them up, and therefore attempts to destroy the old
one and create a new one.

This can be avoided by renaming the resource in the tfstate file, which
is done here using sed.