Allow bootstrap to run fly commands on build
This is an approximation of alphagov/paas-cf@eb1a746. It gives the build
Concourse firewall access to run `fly` commands on the build Concourse and
then removes the firewall rule aftwards. This is required so that we can run
`set-team`, which is used to indicate what environment a Concourse exists
in.

The FIXME has been removed because we're now using a version of Terraform
where the bug has been fixed.