migrate: reattach old volume to the new concourse
In order to keep the concourse history and state, we will reattach the
old concourse volume to the new concourse.

To achieve this we use the script migrate_concourse_volume.sh, which is
divided in two steps:

 * Step 1: Runs before deploy the new concourse is deployed.
   It deletes the old concourse deployed with bosh-init with paas-cf,
   using aws-cli.
   This will deatch the old volume and free the EIP to be used for the
   new concourse.

 * Step 2: Runs after the new concourse is deployed. It will
   1. stop the new concourse.
   2. detach the new volume and delete it.
   3. attach the old volume from the old concourse.
   4. start concourse again.
   5. update the reference to the volume in the Bosh DB directly.

To access the bosh RDS we SSH to concourse itself, that is inside of the
VPC, but we also add the infra_cidrs as allowed IPs to connect to the
RDS instance. We do it with the cidrs to not make it more complicated as
the concourse security group is created with concourse-terraform, that
runs after bosh-terraform. This is OK because it is a temporary commit,
nothing really runs on those Ips and it is still protected by
authentication.