Change name of bosh director to DEPLOY_ENV
Setting the name of bosh director to DEPLOY_ENV will make it more easy
to identify the bosh instance. Also, all the VMs will be tagged with the
tag `director=$DEPLOY_ENV` which would make it easier to operate the
instances using AWS API.

The alternative is add a tag to the deployed VMs, but for that we must
migrate to the new manifest format and include the cloud config
manifest in this repo.