Change deployer-concourse to use m4.xlarge
The `m4.large` instance type that's used by build-concourse isn't sufficient
for our deployer-concourse because it can't cope with the parallel execution
of all test suites. `m4.xlarge` doubles the available CPU and RAM

Per:

- alphagov/paas-cf@81aab6d5b43cd372f14429209cf1e7ce4bbc5523
- https://www.pivotaltracker.com/story/show/134551329

I've parameterised this so that we don't need to unnecessarily increase the
size and cost of the build-concourse.