Read Concourse password for persistent envs
Persistent environments (CI, staging and production) have pre-generated
Concourse passwords that are stored in `pass`. Copy existing functionality
from alphagov/paas-cf that reads these automatically, which prevents us from
accidentally overwriting them. This is a combination of the following
commits:

- alphagov/paas-cf@702070f923cca2c8ef5b2623fc34ea5e99d67fee
- alphagov/paas-cf@1fd588a06145ce641f7cfdb9e2b8a64ce0f71c33

With the addition of `${DEPLOY_ENV}` to the `ci` target, so that it works
for both `build-concourse` and `deployer-concourse` profiles.