Remove datadog-agent from concourse manifest.
This conflicts with the job defined in the runtime config created by
paas-cf. This means that the deploiyment will fail when using a bosh
that has had the runtime-config from paas-cf applied.

This is a temporary fix to unblock the deployment while we unpick the
issue with runtime-config, which will probably need to be migrated to
this repo.