Manage datadog secrets in one script
Currently we manage the datadog secrets by uploading the credentials to
a S3 object in the state bucket, which is read when uploading the
pipelines to populate some variables. These variables would be injected
in the manifests.

In paas-cf the secrets were uploaded adhoc with a make task, and
then read from the pipelines.sh script. There was some logic to ensure
that the file was present if ENABLE_DATADOG=true[1]

But with paas-bootstrap, we need to populate this variables before the
bucket exists, so we have a chicken-egg problem: We cannot upload the
credentials when configuring the pipeline  because the bucket does not
exists, but we need the pipeline to create the bucket.

As a solution, in this commit we move all the logic to manage these
credentials to a script `manage-datadog-secrets.sh`. The script is able
to read from S3 bucket, read from paas-pass if avaiable or environment
variables. It can upload the credentials to S3 too. This way, we check
if the object or bucket are there, and if not, we read from paas-pass.

Note: as appreciation, this problem could be solved by using a centralised
Key-Value store independent of the platform, like credstash or so. But
that implementation is out of the scope.

[1] https://github.com/alphagov/paas-cf/commit/7c3d73d1aecaf48f68b1626ceb2bec8aa9039896