Upload releases from runtime-config
Currently bosh does not automatically upload the missing releases
defined in the runtime config, like it does for normal manifest
deployments[1].

As a workaround, we add an additional script which will iterate the
releases defined in the runtime-config and upload then explicitelly.

We decided to do it in ruby to easily parse the YAML manifest, but shell out
to bosh to keep it as decoupled as possible with the bosh internals.

[1] https://github.com/cloudfoundry/bosh/issues/1364