Move runtime-config management from paas-cf
The runtime-config is global for all the bosh deployments. This causes
a cycled dependency if we keep runtime-config in the paas-cf project:
the first deployment of concourse will not include runtime-config, but
paas-cf will add it later.

It makes more sense to transfer all the code to the paas-bootstrap
repository and manage it from there.

We move all the configuration for collectd, syslog_forwarder and
datadog. Some notes:

 * for clarity, we restructure the files to manifests/runtime-config/addons
   and manifests/runtime-config/addons-meta.

 * collectd: The configuration points to a hardcoded IP. We will
   maintain that for the time being.

 * syslog-forwarder: The syslog target (logstash from logsearch) is
   deployed in paas-cf. We will point to a generated DNS name, which
   will be created/updated during deployment of logsearch.

 * Datadog: the configuration was already in this project, but we change
   the path of the files included. We will update the bosh and concourse
   manifest laters. We get the datadog api from environment variables.

 * We want to be able to toggle some the addons, like collectd and syslog.
   For that, we keep the configuration in multiple files that will be
   conditionally added during the manifest generation.

We add all the related tests included in paas-cf