Add concourse task to render runtime-config
Add a task to render the runtime-config manifest during the bosh
configuration generation. The task will optionally include collectd,
syslog or datadog based on the variables `$ENABLE_COLLECTD_ADDON`,
`$ENABLE_SYSLOG_ADDON` or `ENABLE_DATADOG`

We keep the old name for the datadog one, as the variable is used to enable
other datadog integrations.