Allow SSH from concourse into BOSH
After migrating the bootstrap of bosh and concourse to paas-bootstrap in [1],
we broke the `make <aws_account> ssh_bosh` feature. The reason is that
concourse is not allow to connect to SSH of bosh anymore.

To fix that, we create a new security group bosh_ssh_client in the bosh
terraform, to group the machines that are allowed to SSH to bosh. We
pass the new group ID to the list of SG associated in the concourse
manifest.

[1] https://github.com/alphagov/paas-bootstrap/pull/26