Add an empty releases entry to runtime-config-base
There might be the case that we do not add any addon to the
runtime-config (e.g. when deploying build-concourse in dev without datadog).

In that case the runtime-config will be empty, and we will only use the
runtime-config-base.yml. But bosh fails when running
`bosh update runtime-config` if the key `releases` is missing.

We add an empty entry `releases:` to the runtime-config-base.yml to
workaround that:

```
Error 40001: Required property 'releases' was not specified in object ({"addons"=>[]})
```

We consider the error is intended, so we will not open an issue
upstream.