Switch to Docker
We will use official concourse docker image, with the recommended way of
using compose to deploy it. See [1] for more details. We generate database
credentials in the environment script and we need to use hex password for
postgres, because if it contains any /, then atc parses that wrong and it
takes a part of the password as port and fails to connect.

For the vagrant box that will run docker compose we can use any image. We
have chosen to use default official ubuntu 16.04 release. We bump box to
m4.large, as it's a bit faster and cheaper than m3 and future proof. The
new image is EBS based, so we don't need local storage. We increase default
disk size to 50G to accommodate all the docker images and resources needed
by the pipelines.

Because we use default ubuntu, we need to generate a SSH key for this box
so that Vagrant can run post deploy scripts and so that we can ssh to the
host in case we need to do any debug or other operations. We generate the
key in the current directory (not in vagrant, as that would rsync the
private key to the box) and add it to gitignore. We clean up this key upon
VM termination in self terminate task and in the bootstrap-destroy makefile
target.

Fly update might not always work as you expect, so we add a note about that.

Similarly, because docker compose needs to have stdout when starting
containers, we have to sleep for a while (currently 1m) in the provision
scripts. Creating tunnel can fail if the timing is not right and concourse
is still starting up. Document that this can be fixed by re-running make
target.

Simplify post-deploy script by symlinking all necessary files to vagrant
directory, which gets rsynced to the host on start. This means we don't
need to clone this repository to get scripts - saves time and removes one
dependency.

Run docker compose as a daemon and wait for concourse to start up
in a loop with 3 minute timeout. Update readme to reflect this.

[1] https://concourse.ci/docker-repository.html