Fix bootstrap pipelines
In my enthusiasm I have forgotten to define outputs for mostly TF tasks.
That is, I let TF run directly on file that is in the input directory and
update it in place. That works fine, but if you try to 'put' that file
later, that is a different container and it doesn't have access to that
file. So you need to explicitly define it as an output. Fix the pipelines to
do so.