Move Bosh RDS maintenance window to 6-7am
RDS maintenance windows can cause short and in exceptional cases longer
periods of downtime.

We have more people available to deal with incidents when it's closer
working hours, but we still want to have the maintenance window before
the working day when tenants are likely to be using the PaaS more.