Create and use ACM cert for concourse.
This hooks in the script to the pipeline so that concourse will use the
ACM provided cert.

This has replaced the previous task which generated a self-signed
certificate. With ACM, we also no longer need a concourse resource to
track this cert because it's all handled on the ACM side, and accessed
with a terraform data resource.