Add a btrfs process count metric on concourse
This is to allow us to add a datadog monitor that fires when a concourse
node goes wild and needs shaking.

The empty `init_config` block is present because datadog requires it and
will error if it is not found. Normally it would contain configuration
information for the integration, but the `process` plugin does not
require any.