Allow Bosh RDS instance to auto_minor_version_upgrade
So that we automatically get upgraded to newer point releases when there
are security issues etc.

This includes updating the engine_version to only specify the major
version so that terraform won't attempt to downgrade instances that have
been upgraded. It also means that newly created instances will get the
latest point release.