Increase max number of retries
We're implementing a functionality that'll kick-off the
dev `create-cloudfoundry` pipeline every morning. Starting many
environments at roughly the same time, will cause increased amount of
calls made towards AWS.

Increasing the number of max retries, should help us a little with the
potential issue of being caught out by the limit established upstream.

It is set to 16, as previously (a while ago), we had it set to 8. We
have more people on the team and 16 felt like a safe bet to play with.

I have ran the bootstrap on my `dev` environment, which didn't cause any
issues with regards to the change.