Implement commit signature verification
We intend to push the create/destroy pipelines to the Concourses
created by paas-bootstrap (although only the 'create' pipeline to persistent
environtments) and eventually have these pipelines auto-triggered,
therefore we want to have an extra guard in place to make sure the change
has been authorised by a member of the team.

The implementation was copied from paas-cf[1]

If we reuse the merge signing script anywhere else we should consider extracting
it into an external helper, such as a ruby gem.

[1] https://github.com/alphagov/paas-cf/tree/811da924b62da9de28e6bd0d7d7c385f768982ef