Upload pipelines to target Concourse
We want to run the bootstrap pipelines (create and destroy) on the
Concourses in our deployed environments. This will allow us to create
a bootstrap Concourse only once and run the pipelines in the created
Concourse from then on.

First of all, to support only creating a bootstrap once, we need to
make the pipeline self-update. This cannot happen in the bootstrap
Concourse because uploading the pipelines involve determining the
Concourse ATC password, which is derived from our AWS user IDs, and
therefore specific to the operator who created the bootstrap. The self-
update also has to come after the initialisation of the bucket, because
otherwise it cannot access the pipeline trigger. We have to split the
initialisation of S3 files into a separate job, as Concourse loads all
tasks for a job into memory, then runs them, so it would not get updated
if it stayed in the same job.

In order to know which Concourse we are targetting and differentiate
the the deployer and build Concourses we had to introduce two environment
variables:

`TARGET_CONCOURSE` is the Concourse which pipelines are being uploaded to,
whereas `CONCOURSE_TYPE` is strictly there to differentiate the build
and deployer Concourses, as they have different IAM roles.

We use the target Concourse for two main purposes: firstly, to make sure
we only upload the creation pipelines to our Concourses, as the Concourse
cannot and should not destroy itself; and secondly, to enable us to use
an external host name for Bosh when communicating from the bootstrap, and
an internal one when operating from the Concourse deployed by the bootstrap.

We also stop setting VAGRANT_IP when running pipeline update; this always runs
on the build/deployer concourse, where it is unnecessary and adds an IP address
that no longer belongs to us to the security group.