Access BOSH via public IP
Previously we had boostrap concourse accessing bosh via its public IP and
deployer/build concourse via its internal IP.

However this involved a change to the bosh manifest and that meant that
bosh-init re-deploys bosh when the pipeline runs on the persistent concourse
for the first time, which is painful to say the least.

We now access bosh via public IP in all cases, which we access using the EC2
metadata API from within the pipeline, rather than when we provision the
bootstrap concourse with vagrant.

For the duration of the pipeline run the security group rules allow whichever
concourse is running the pipeline to access ports on the BOSH director machine
via its public IP.