Use fly --non-interactive argument
The new fly cli has a problem with things being piped into it. This
causes issues whenever we're trying to use an interactive way.
Thankfully, there is a way to disable the interactive mode which helps
us move forward. This does however mean, we're using username and
password inline when performing a login.