Fix typo in SSH key generation
Checking in the wrong directory/resource caused the SSH keys to be recreated
each time these two tasks were run.

I noticed this problem working while testing the instructions to access BOSH
without Concourse, after I had run the `create-bosh-concourse` pipeline to
test the Concourse upgrade. Although Terraform upload the new keypair to
AWS, it doesn't take effect if `bosh-init` doesn't recreate the BOSH VM.

I've tested the persistent environments and it looks like the key in the
state bucket still works at the moment, probably because we upgraded the
BOSH VM the last time this was run. However we need to fix this before the
Concourse upgrade is applied to those environments.