Enable IPv6 support in Concourse VM.
IPv6 is disabled by default in the stemcell. A control has been added to
bosh-agent to re-enable this[1]. This has the effect of removing the
"ipv6.disable=1" flag from the kernel commandline.

We hit a bug in gpg[2] where it will fail to connect to a keyserver if
IPv6 support is not enabled in the kernel. The error returned is:

gpg: keyserver receive failed: Address family not supported by protocol

We've worked around this by explicitly specifying a keyserver that only
has IPv4 addresses in 7866ee6.

Enabling IPv6 support in the kernel avoids the need for this workaround.
It's not necessary for the IPv6 stack to be configured, it merely needs
to be present.

[1]https://github.com/cloudfoundry/bosh-agent/pull/129
[2]https://dev.gnupg.org/T3331