Upgrade stemcell for CVEs
To fix the following:

- https://www.cloudfoundry.org/usn-3424-1/
- https://www.cloudfoundry.org/usn-3434-1/
- https://www.cloudfoundry.org/usn-3444-2/
- https://www.cloudfoundry.org/usn-3454-1/
- https://www.cloudfoundry.org/usn-3441-1/
- https://www.cloudfoundry.org/usn-3432-1/

The series has been updated to match alphagov/paas-cf.

We're using 3445.16 instead of the recommended minimum of 3445.15 because
.15 doesn't have a "light" stemcell published. The same URL redirects to the
"fat" version, which needs to be mounted as an EBS volume on the host
running `bosh-init`, which doesn't work from a container:

    Starting registry... Finished (00:00:00)
    Uploading stemcell 'bosh-aws-xen-hvm-ubuntu-trusty-go_agent/3445.15'... Failed (00:01:23)
    Stopping registry... Finished (00:00:00)
    Cleaning up rendered CPI jobs... Finished (00:00:00)

    Command 'deploy' failed:
      creating stemcell (bosh-aws-xen-hvm-ubuntu-trusty-go_agent 3445.15):
        CPI 'create_stemcell' method responded with error: CmdError{"type":"Bosh::Clouds::CloudError","message":"Cannot find EBS volume on current instance","ok_to_retry":false}

I've been talking to some people on #bosh in the CF Slack about improving the
error message because at first I thought it was something I'd broken in my dev
environment when messing with manifests to test AZ failover.