Script to execute Docker-based bosh-shell
This pulls credentials and Bosh IP from S3, then executes the
governmentpaas/bosh-shell Docker image, passing in the secrets as
environment variables.

All of this assumes that the image has a ENTRYPOINT that sets up Bosh
and drops the user into a useful shell.

We have exported these variables so that Docker can propagate them into
the container without them being readable via the `ps` command on the
host machine.