Use github oAuth for logging into concourse
This commit adds a list of github users to add into the fly
team creation command. These are seperated into Prod and CI users for CI
environments both user lists are used. For staging and prod only the
prod list is used.

There is also an additional make target to upload the github auth
credentials to persistent environments. Basic auth is retained as a
fallback and for script usage.