Add script for SSHing to Bosh
This was taken from paas-cf and adapted to SSH directly without going
via Concourse. This is because the security group for the Bosh VM now
allows access from our office IPs.

It's been necessary to silence SC2029[1] in this script as it was
complaining about the word being evaluated on the remote side, not the
client side. This doesn't seem to apply here as the variable is actually
the remote IP to connect to, and therefore has to be evaluated
client-side. It's probably that it's the user part of this string that
would be potentially vulnerable, and that shellcheck is unable to
determine which part of the string has the variable in it.

[1]https://github.com/koalaman/shellcheck/wiki/SC2029