rubocop: Ignore block lengths in spec dirs
There was an upstream config change that describes why we don't want to
check block lengths in tests:

- alphagov/govuk-lint@195e1d6934e610efed35151cfe90b01b805d7581

However it wasn't working for us because our `spec` directories are nested
within other directories. We can workaround it with a local config for now,
but it can be removed when the upstream PR is merged and released.