Upgrade Concourse to 3.5.0
This is the first step of the Concourse upgrade process from 3.4.1 -> 3.8.0.

This change only updates the Concourse BOSH release version, which enables an
upgrade path to swap out the bundled PostgreSQL job with the one using the
official Postgres BOSH release.

Following the upgrade notes written in https://concourse.ci/downloads.html#v350.