Set SYSTEM_DNS_ZONE_ID in the Makefile for the different envs
Instead of passing the root DNS zone name and looking up the zone id in the
create certificate script just use the values we already have as Terraform
variables. To avoid duplicating configuration we move the zone ids to the
Makefile and pass them to Terraform and the cert script.