Set serial:true for all Concourse jobs
While working on a story I noticed that it was possible to run two parallel
instances of the `bosh-terraform` job, both of which were trying to modify
the same resources.

Apply `serial: true` to all other jobs in this pipeline so that they only
execute once at a given time otherwise there will be race conditions and
conflicts.