terraform: Stop using -state-out argument
We have noticed that this can result in an empty (zero byte) state file
being stored in the resource is Terraform is terminated during operation.
This can happen if you omit the `-auto-approve=true` argument in recent
Terraform versions and subsequently abort the job in Concourse.

We still have to maintain two files because Concourse requires the output
directory to be separate from the input directory. We solve this by copying
the input to the output before running Terraform.