Fix pinning of paas-bootstrap resource in pipeline
The version of this resource used in the s3init job wasn't pinned to use
the same version as in the in init-bucket job. This is inconsistent with
other pipelines, and causes mistakes when attempting to run the pipeline
with a specific version by disabling other versions in the UI etc.