Move Github OAuth config to the manifest.
Currently the config for the main team is being lost every time atc
restarts. This is because we're not supposed to use `fly set-team` to
configure the "main" team[1]. This is also covered in the docs[2].

This migrates this config into the manifest, so that it will persist
across reboots etc.

[1] https://github.com/concourse/concourse/issues/1963
[2] http://concourse.ci/teams.html#main-team