Generate ACM certificates also in us-east-1
There is a restriction with ACM certificates when using CloudFront,
the certificate must be generated in the eu-east-1 region[1]. But
we still need it in the region where the ELB use it. Otherwise
it errors with:

   InvalidViewerCertificate: The specified SSL certificate doesn't exist,
   isn't in us-east-1 region, isn't valid, or doesn't include a valid
   certificate chain.

To solve this issue, we generate the certificate in the current
region AND also in us-west-1.

[1] https://docs.aws.amazon.com/acm/latest/userguide/acm-services.html