Stop creating ACM certs in us-east-1
These were created with the expectation that they'd be used for the
product page and tech docs CloudFront deployments. We've since decided
that we're instead going to get rid of these CloudFront instances and
route the traffic directly to the paas. These certs are therefore no
longer needed.

This partially reverts 61d377f