Add script for running arbitrary Bosh commands
Sometimes we may want to execute Bosh commands and pipe the output
to other processes. For this, the existing bosh-cli `make` target
was not suitable, as it is an interactive process.

This script makes use of a new Docker image[1], which merely passes
all `docker run` arguments to the `bosh` command. The script does
connect STDIO, so `bosh ssh` is supported as well.

[1] https://github.com/alphagov/paas-docker-cloudfoundry-tools/tree/master/bosh