Disable use of Client TLS with Logit
We have been unable to manually configure Client TLS using `rsyslog`
on the VMs. We are not sure why, as rsyslog 8.22 (the same as used on
the VMs) could do it when run on a Vagrant VM [1].

This commit disables using Client TLS to authenticate clients, in
anticipation that we will rely on IP Whitelisting instead.

[1] https://www.pivotaltracker.com/n/projects/1275640/stories/156996834/comments/189490255