Revert "Implement mutual TLS with Logit"
Rationalle[1]:

Mutual TLS carries some risks/complications around what happens when the
client certs expire, and what the process for rotating the certs is.
These certs have to be configured via a support ticket, and we're not
clear what the response time is there, or whether we could have 2 certs
live at once etc.

It's also not clear that there's really a need for it. With an open
logstash endpoint, the worst that can happen is that someone discovers
the endpoint and starts sending junk to it. I think the likelyhood of
this is very low as there's not much to be gained by doing it.

For the staging and prod stacks, we can enable IP whitelisting for them,
but we can pick that up in a follow-up story. For dev I think it's fine
to leave it open.

[1] https://www.pivotaltracker.com/story/show/156996834/comments/189760131