Pass the feature flag env vars to the pipelines.sh/self-update-pipeline.sh
This change will allow you to set feature flags without SELF_UPDATE_PIPELINE=false.

The affected flags:
 - ENABLE_DATADOG
 - ENABLE_GITHUB

Currently when we set a feature flag we have to disable self-updating the
pipeline as it resets these flags.
Most of the time this is inconvenient as you have to update the pipelines
manually every time you make a change.
Also testing the self update process itself with feature flags enabled is
currently not possible.