Remove runtime-config.
Now that there's nothing left in it, we can remove most of the code to
do with generating and applying the runtime config.

The only thing remaining is a task to apply a blank runtime config. This
is necessary to ensure that the runtime-config gets cleared everywhere.
Once this has happened, we can remove this task as well.