Add `make` targets for London environments
The new system and apps domains for London environments sit under the
domains of their Ireland counterparts, so we hard code them as
environment variables. Here we create new `make` targets which pin both
the AWS account and the environment, to be absolutely sure everything is
set correctly before deploying.