Don't assume what the AWS account is
This has been removed because not knowing what account you are meant
to be targetting should result in fast failure, to avoid doing something
destructive.

This was actually causing a subtle bug in the Vagrant scripting as well.
Because it wasn't being passed to the post-deploy script environment,
it actually had the default value `dev` set during those scripts.
Fortunately, none of the post-deploy scripts required this variable.