Self-update pipelines using Makefile target
Previously the Makefile target names happened to coincide with the
`AWS_ACCOUNT` environment variable. However, we have introduced new
system and apps domains for staging and prod environments in the London
region. These domains are not like-for-like in terms of where they sit
in the subdomain heirarchy, therefore they get hardcoded as environment
variables in the `make` target. Because of this, the simplest solution
is to add new targets for our London environments and use the new
`MAKEFILE_ENV_TARGET` environment variable to handle the self-updating of
pipelines.