Don't check DEPLOY_ENV before it's defined
The prod, staging etc env targets now define the DEPLOY_ENV for these
environments, however, the check-env-vars target was being invoked
before this definition leadint to errors when your shell didn't have
this env ver defined, even though the shell defined one would be
ignored.

This updates the checks to be invoked only at the point that they're
needed to prevent this issue.