Bump Bosh to 266.6.0
We have encountered an issue [1] of a process putting a VM under load that causes
monit to drop the connections. This is fixed in the 266.x series, therefore we
should upgrade to that.

Version 266.6.0 may have fixed another issue as well[2].

[1] cloudfoundry/bosh#1754
[2] cloudfoundry/bosh#1990

For reference, the Precompiled releases are available in the public bucket:
https://s3.amazonaws.com/bosh-compiled-release-tarballs

Version IDs can be extracted using something like:

```bash
aws s3api list-object-versions --bucket bosh-compiled-release-tarballs --prefix bosh-266.6
```