Update the concourse stemcell to force recreate
When enabling the syslog-forwarder with the syslog-release,
that release basically copies a bunch of files into
/etc/syslog.d

But when deleting the syslog-release these files never get deleted,
so we need to find a way to get rid of them to stop concourse
shipping logs.

The simplest way is to leverage the bosh pattern of disposable servers
and simply change the stemcell version, forcing bosh to recreate
the VM. This will activelly wipe all these files as a new
VM will be created.