v3.6.0 add creds and cred stealing job
Used YAML templating so that we are less verbose

cf-deployment v2 -> v3.6.0 upgrade requires some more PKI which we have
dutifully added.

We were deploying bosh-dns as part of cf-deployment, but in v3.0.0 this
is expected to be done in bosh runtime-config and taken out of
cf-deployment. In order to do this as smoothly (in terms of downtime) as
possible we would like to re-use our old PKI for bosh-dns.

There is a temporary (read FIXME) task in the pipeline which attempts to
borrow the existing PKI. After v3.6.0 is everywhere we can come back and
make this nice. Currently this task does actually do the right thing for
all deployment cases (new deploy, existing upgrade deploy, existing noop
deploy)

author:     @jpluscplum
amanuensis: @tlwr