[#159730005] add generate-environment-variables task
Previously we had a task `steal-cf-vars` which copied bosh-dns related
pki material from cf-vars-store to bosh-runtime-config-vars-store

Now that bosh-dns has been rolled out everywhere (that we care about:
staging, prod) we can get rid of the copying task. The functionality of
this task is now in the `set-runtime-configs` task, which can now
generate the variables if they do not exist.

Expected behaviour is that new environments have the vars generated, and
existing environments use the existing vars.

We use separate `bosh interpolate` and `update-runtime-config` commands
because we want to engage bosh's `--var-errs` protection, but
`uptime-runtime-config` doesn't respect this flag.