Remove Environment specific team name
The concept of teams has changed significantly in concourse 4.0.0 and
up. Having 2 teams with the same username breaks the migration as a
result.

Then environment team name was added as a caution after an accidental
running of the delete pipelines was made. We have since safe guarded
against this by no longer pushing the delete pipelines to persistent
environments.