Consume upstream component versions
A semi-official upstream exists [1] which ties together the suggested
versions of concourse, garden-runc and postgres, released on a similar
cadence to concourse itself. We consume this in the CI pipeline, as a
resource, and pin to a release version (or version prefix) we wish to
be running.

We don't trigger from releases of this resource so that we maintain
control over when our workers are potentially bumped.

As we're using `spruce` to output bosh manifests, we need to use the
`((!foo))` syntax mentioned here [2] to avoid spruce thinking it should
be interpolating the variables at manifest generation time. This is a
stepping stone on the route towards using `bosh interpolate` instead of
`spruce` - it feels wrong to add more responsiblities into `spruce`'s
manifest generation in this repo.

[1] https://github.com/concourse/concourse-bosh-deployment
[2] https://github.com/geofffranks/spruce/blob/master/doc/integrating-with-credhub.md