Bump Concourse stemcell to latest
The Concourse 3.10.0 release notes [1] suggest that the latest possible
btrfs kernel version is strongly desireable, for stability and
performance reasons.  This commit bumps us to a couple of versions
behind the latest Trusty version.

[ That "couple of versions" were chosen to allow us some slack to test
pipeline / Concourse-flow upgrades of minor stemcell releases, but that
dropped out of scope of the story being worked on. Given the choice of
redoing all the testing on the very latest Trusty stemcell, or to merge
this work, it was felt more appropriate to stick with the slightly older
stemcells which had been more rigourously testing in our environment and
get this merged. ]

[1] https://concourse-ci.org/download.html#v3100
[2] https://www.pivotaltracker.com/story/show/158180318