Use upstream docker-compose file for bootstrap
Concourse upstream publishes a docker-compose file which uses their new
`quickstart` command to auto-wire a worker and atc together. This commit
uses that upstream file (vendorered, as a couple of changes need to be
accepted upstream before we can use it [1]).

To use it, we needed to upgrade to Bionic as our bootstrap substrate for
a later version of docker-compose, which introduced a subtle cloud-init
boot-time timing problem, where Vagrant can run our commands (`apt-get
update`) before the apt config files are pointing at the correct repos.

[1] https://github.com/concourse/docs/pull/112